+++ packages/open-api-client/CHANGELOG.md
@@ -2,6 +2,12 @@
 All notable changes to this project will be documented in this file. See [conventional commits](https://www.conventionalcommits.org/) for commit guidelines.
 
 - - -
+## open-api-client-v0.1.4 - 2025-10-30
+#### Bug Fixes
+- (**open-api-client**) quote property keys - (51fe91a) - *brad-jones*
+
+- - -
+
 ## open-api-client-v0.1.3 - 2025-10-30
 #### Bug Fixes
 - (**formatter**) forgot to await Deno.readFile in downloadWasmFormatter - (7540ce2) - *brad-jones*